UpdateEventLevelArgs.Levelの型をInt64からUintに変更

+++ interfaces/repository/event_impl.go
@@ -102,7 +102,7 @@ func (repo *EventRepository) UpdateEventLevel(eventID uuid.UUID, arg *repository
 	err := repo.h.Transaction(func(tx database.SQLHandler) error {
 		if elv, err := repo.getEventLevelByID(eventID); err != nil {
 			return convertError(err)
-		} else if uint(elv.Level) == uint(arg.Level.Int64) {
+		} else if uint(elv.Level) == uint(arg.Level.Byte) {
 			return nil // updateする必要がないのでここでcommitする
 		}
 		/*} else if elv.Level == arg.Level {
